add trackValueSelector and comparator option
